Spread responsibility when retrieving asset
